pyproject.toml
[tool.poetry]
name = "telethon-script"
version = "0.1.0"
description = ""
authors = ["A-V-tor <avtorca4@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.33.1"
flask-qrcode = "^3.1.0"
quart = {extras = ["dotenv"], version = "^0.19.4"}
pydantic = "^2.5.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
